Split fan-out routing module
Move lane workers, filter logic, compression training, and fallback dispatch into focused fan_out submodules.

Keep FanOutSend and Submitter as the coordination layer.
